Исправлены тесты на класс Lion
